Hadoop and Java movie recommender system, along with C++ pipes version to display differences
This is a movie recommender system with 2 map-reduce phases, written in
both Java/Hadoop and C++. The idea is that the user gives a list of
movies and their personal rating (1-5 stars), then in the first map
reduce phase a confidence rating is built between a giant list of movies
and reviews by reviewers. After that, a second map-reduce is built by
weighting each movie based on the confidence rating obtained from the
first, then output the n-top movies the user requests based on what they
haven't seen. Example run given in "HadoopRun.txt"